refactor(routing): move generateTlsSummary back to VirtualClusterModel
generateTlsSummary was moved onto UpstreamClusterModel during a previous
refactoring, but it is not specific to upstream clusters — it also serves
the gateway's downstream TLS summary. VirtualClusterModel is the right
home since it owns both usages. The method is package-private so
VirtualClusterModelTest can exercise it directly. The tlsSummary()
instance method is removed from UpstreamClusterModel; VirtualClusterModel
now calls generateTlsSummary(cluster.tls()) directly.

Tests move accordingly: the generateTlsSummary cases leave
UpstreamClusterModelTest and join VirtualClusterModelTest.
