+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/model/VirtualClusterModelTest.java
@@ -21,6 +21,7 @@ import io.kroxylicious.proxy.config.NamedFilterDefinition;
 import io.kroxylicious.proxy.config.PluginFactory;
 import io.kroxylicious.proxy.config.PluginFactoryRegistry;
 import io.kroxylicious.proxy.config.TargetCluster;
+import io.kroxylicious.proxy.config.tls.AllowDeny;
 import io.kroxylicious.proxy.config.tls.Tls;
 import io.kroxylicious.proxy.config.tls.TlsCredentialSupplierConfig;
 import io.kroxylicious.proxy.filter.FilterFactory;
@@ -292,6 +293,47 @@ class VirtualClusterModelTest {
                 };
             }
         };
+    }
+
+    // generateTlsSummary()
+
+    @Test
+    void generateTlsSummaryReturnsEmptyStringForNoTls() {
+        assertThat(VirtualClusterModel.generateTlsSummary(Optional.empty())).isEmpty();
+    }
+
+    @Test
+    void generateTlsSummaryIncludesTlsMarkerWhenTlsPresent() {
+        var summary = VirtualClusterModel.generateTlsSummary(Optional.of(new Tls(null, null, null, null, null)));
+        assertThat(summary).contains("(TLS:");
+    }
+
+    @Test
+    void generateTlsSummaryIncludesAllowedCipherSuites() {
+        var tls = new Tls(null, null, new AllowDeny<>(List.of("TLS_AES_256_GCM_SHA384"), null), null, null);
+        var summary = VirtualClusterModel.generateTlsSummary(Optional.of(tls));
+        assertThat(summary).contains("Allowed Ciphers").contains("TLS_AES_256_GCM_SHA384");
+    }
+
+    @Test
+    void generateTlsSummaryIncludesDeniedCipherSuites() {
+        var tls = new Tls(null, null, new AllowDeny<>(null, Set.of("TLS_RSA_WITH_NULL_MD5")), null, null);
+        var summary = VirtualClusterModel.generateTlsSummary(Optional.of(tls));
+        assertThat(summary).contains("Denied Ciphers").contains("TLS_RSA_WITH_NULL_MD5");
+    }
+
+    @Test
+    void generateTlsSummaryIncludesAllowedProtocols() {
+        var tls = new Tls(null, null, null, new AllowDeny<>(List.of("TLSv1.3"), null), null);
+        var summary = VirtualClusterModel.generateTlsSummary(Optional.of(tls));
+        assertThat(summary).contains("Allowed Protocols").contains("TLSv1.3");
+    }
+
+    @Test
+    void generateTlsSummaryIncludesDeniedProtocols() {
+        var tls = new Tls(null, null, null, new AllowDeny<>(null, Set.of("TLSv1.1")), null);
+        var summary = VirtualClusterModel.generateTlsSummary(Optional.of(tls));
+        assertThat(summary).contains("Denied Protocols").contains("TLSv1.1");
     }
 
     private static PluginFactoryRegistry pluginFactoryRegistry() {
